The gegy is trying something

+++ src/main/java/net/timeless/jurassicraft/common/entity/base/EntityDinosaur.java
@@ -231,7 +231,7 @@ public class EntityDinosaur extends EntityAICreature implements IEntityAdditiona
         adjustHitbox();
     }
 
-    public double getDinosaurAge()
+    public int getDinosaurAge()
     {
         return dinosaurAge;
     }
